+++ contracts/Portal/Portal.sol
@@ -45,7 +45,7 @@ import {StakeModule} from "./modules/StakeModule/StakeModule.sol";
  * * * Pausing gETH, pausing Portal, releasing prisoned operators,and seting a governance fee.
  * * Push end of the version management logic via pull->push.
  * * * approveProposal changes the package version or allows specified middleware.
- * * * pushUpgrade creates a contract upgrade proposal on the package, and requires package owner to approve it.
+ * * * pushUpgrade creates a package type proposal on the package to upgrade the contract, and requires package owner to approve it.
  *
  * @dev authentication:
  * * GeodeModule has OnlyGovernance, OnlySenate and OnlyController checks with modifiers.
@@ -212,19 +212,18 @@ contract Portal is IPortal, StakeModule, GeodeModule {
 
   function pushUpgrade(
     uint256 packageType
-  ) external virtual override whenNotPaused nonReentrant returns (bytes memory versionName) {
+  ) external virtual override whenNotPaused nonReentrant returns (uint256 id) {
     require(
       packageType > ID_TYPE.LIMIT_MIN_PACKAGE && packageType < ID_TYPE.LIMIT_MAX_PACKAGE,
       "PORTAL:invalid package type"
     );
 
     uint256 currentPackageVersion = STAKE.packages[packageType];
-    versionName = DATASTORE.readBytes(currentPackageVersion, "NAME");
 
-    uint256 id = IGeodeModule(msg.sender).propose(
+    id = IGeodeModule(msg.sender).propose(
       DATASTORE.readAddress(currentPackageVersion, "CONTROLLER"),
-      ID_TYPE.CONTRACT_UPGRADE,
-      versionName,
+      packageType,
+      DATASTORE.readBytes(currentPackageVersion, "NAME"),
       GML.MAX_PROPOSAL_DURATION
     );