+++ contracts/Portal/packages/LiquidityPool.sol
@@ -207,8 +207,8 @@ contract LiquidityPool is ILiquidityPool, LiquidityModule, GeodeModule {
     require(!(getPortal().isolationMode()), "LPP:Portal is isolated");
     require(getProposedVersion() != getContractVersion(), "LPP:no upgrades");
 
-    bytes memory versionName = getPortal().pushUpgrade(GEODE.PACKAGE_TYPE);
-    approveProposal(DSML.generateId(versionName, ID_TYPE.CONTRACT_UPGRADE));
+    uint256 id = getPortal().pushUpgrade(GEODE.PACKAGE_TYPE);
+    approveProposal(id);
   }
 
   /**