+++ audits/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Curly brackets are added to lacking if statements
 - `\_decreaseWalletBalance` function call bring to top to save gas on `proposeStake` function
 - Improve `fallbackOperator` logic by making fallback treshold settable
+- Package update logic (Limited UUPS) changed to use package type as proposal type instead of `CONTRACT_UPGRADE (TYPE 2)`
 
 ### Fixed